pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scrape-to-md"
version = "0.1.0"
description = "Simple URL scraper that outputs markdown files"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "trafilatura>=1.6.0",
    "yt-dlp>=2024.1.0",
    "youtube-transcript-api>=0.6.0",
    "docling>=2.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.9.0",
]

[project.scripts]
scrape_to_md = "scrape_to_md.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
